# perf: Improve `@babel/generator` performance