--- a/packages/babel-generator/src/buffer.ts
+++ b/packages/babel-generator/src/buffer.ts
@@ -29,8 +29,13 @@ type QueueItem = {
 };
 
 export default class Buffer {
-  constructor(map?: SourceMap | null) {
+  constructor(map: SourceMap | null, indentChar: number) {
     this._map = map;
+    this._indentChar = indentChar;
+
+    for (let i = 0; i < 64; i++) {
+      this._fastIndentations.push(String.fromCharCode(indentChar).repeat(i));
+    }
 
     this._allocQueue();
   }
@@ -43,6 +48,8 @@ export default class Buffer {
   _queue: QueueItem[] = [];
   _queueCursor = 0;
   _canMarkIdName = true;
+  _indentChar = 0;
+  _fastIndentations: string[] = [];
 
   _position = {
     line: 1,
@@ -187,8 +194,9 @@ export default class Buffer {
   /**
    * Same as queue, but this indentation will never have a sourcemap marker.
    */
-  queueIndentation(char: number, repeat: number): void {
-    this._pushQueue(char, repeat, undefined, undefined, undefined);
+  queueIndentation(repeat: number): void {
+    if (repeat === 0) return;
+    this._pushQueue(-1, repeat, undefined, undefined, undefined);
   }
 
   _flush(): void {
@@ -208,10 +216,21 @@ export default class Buffer {
   ): void {
     this._last = char;
 
-    this._str +=
-      repeat > 1
-        ? String.fromCharCode(char).repeat(repeat)
-        : String.fromCharCode(char);
+    if (char === -1) {
+      const fastIndentation = this._fastIndentations[repeat];
+      if (fastIndentation !== undefined) {
+        this._str += fastIndentation;
+      } else {
+        char = this._indentChar;
+      }
+    }
+
+    if (char !== -1) {
+      this._str +=
+        repeat > 1
+          ? String.fromCharCode(char).repeat(repeat)
+          : String.fromCharCode(char);
+    }
 
     if (char !== charcodes.lineFeed) {
       this._mark(