<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | Closes https://github.com/babel/babel/pull/16051 <!-- remove the (`) quotes and write "Fixes" before the number to link the issues -->
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
Time consumption reduced by ~15%
```
PS F:\babel\benchmark\babel-generator\real-case> node .\jquery.mjs
current 1 jquery 3.6: 165 ops/sec ±0.61% (6.056ms)
current 4 jquery 3.6: 39.67 ops/sec ±0.98% (25ms)
current 16 jquery 3.6: 9.31 ops/sec ±3.4% (107ms)
current 64 jquery 3.6: 2.3 ops/sec ±1.37% (435ms)
baseline 1 jquery 3.6: 135 ops/sec ±0.71% (7.41ms)
baseline 4 jquery 3.6: 32.78 ops/sec ±2.38% (31ms)
baseline 16 jquery 3.6: 8.11 ops/sec ±0.7% (123ms)
baseline 64 jquery 3.6: 1.95 ops/sec ±1.05% (513ms)
```